Update: Add: main and game logic files
Change main for functional draft of game
Add game logic file with classes, not yet OOP

We are missing class game, loop for turns, winner class and details
