+++ README.md
@@ -1,6 +1,13 @@
 # OOP Tic Tac Toe
 
-> This game is built using Object Oriented Programming concepts. Particularly, classes and access to their attributes. It involves a couple of players, a simple board, checking for victory in a game loop
+> This game is built using Object Oriented Programming concepts. Particularly,
+> classes and access to their attributes. It involves a couple of players, a
+> simple board, checking for victory in a game loop
+
+The main goal of this project is to put into practice concepts learned in Object
+Oriented Programming main concepts of Object Oriented Programming. Particularly,
+classes and access to their attributes by building this project. It involves a
+couple of players, a simple board, checking for victory in a game loop
 
 ![screenshot](./app_screenshot.png)
 
@@ -63,4 +70,4 @@ Give a ⭐️ if you like this project!
 
 ## 📝 License
 
-This project is [MIT](LICENSE.txt) licensed.
+This project is [MIT](LICENSE.md) licensed.
